pyproject.toml
[project]
name = "DisplayPad"
version = "1.0.3"
authors = [
  { name="Sytxlabs", email="info@sytxlabs.eu" },
]
description = "This library allows you to customize your own Mountain DisplayPad by assigning each button its own custom function, image or color."
readme = "README.md"
requires-python = ">=3.10"
classifiers = []
dependencies = ["hidapi==0.14.0.post2", "pyee==12.0.0", "Pillow==11.0.0"]

[project.urls]
Homepage = "https://sytxlabs.eu"
GitHub = "https://github.com/SytxLabs/DisplayPad"

[project.license]
text = "MIT license"

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"